feat(bot): foundation for @gracechords_bot refactor — migrations + pure PDF renderer
- Add telegram_user_id / telegram_linked_at columns on users with a partial
  index for the bot's hot-path Telegram → user lookup.
- Add songs.search_titles[] plus a trigger that auto-populates variants
  (parenthetical, slash-split, normalized) on insert/update, a pg_trgm GIN
  index, a backfill, and a bot_search_songs(q) RPC the bot uses.
- Split src/utils/pdf_mvp/index.js into a DOM-free pure.js (jsPDF layout
  algorithm + new renderSingleSongPdfBuffer / renderMultiSongPdfBuffer
  returning Uint8Array) and a thin browser shim that keeps every existing
  public export's signature. SongViewPage / SetlistPage / songbook callers
  unchanged. All 9 pdf_mvp.*.test.js still pass.
